% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_simplex_algorithm.R
\name{construct_dual_feasible_tableau}
\alias{construct_dual_feasible_tableau}
\title{construct_dual_feasible_tableau}
\usage{
construct_dual_feasible_tableau(A, b, c, sense)
}
\arguments{
\item{A}{\code{matrix}, coefficient matrix of the LP}

\item{b}{\code{vector}, RHS}

\item{c}{\code{vector}, coefficients of the objective function}

\item{sense}{\code{scalar}, max = 1, min = -1 (default = 1)}
}
\value{
\code{matrix}, initial (dual feasible) simplex tableau
}
\description{
construct_dual_feasible_tableau
}
