% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_simplex_algorithm.R
\name{dual_simplex}
\alias{dual_simplex}
\title{dual_simplex}
\usage{
dual_simplex(tableau, max_iter = 100)
}
\arguments{
\item{tableau}{\code{matrix}, a simplex tableau}

\item{max_iter}{\code{scalar}, maximum number of iterations}
}
\value{
(primal and dual) optimal simplex tableau or stops execution in case of an unbounded problem
}
\description{
dual_simplex
}
