% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviced_primal_simplex_agorithm.R
\name{get_arg_min_conditional}
\alias{get_arg_min_conditional}
\title{get_arg_min_conditional}
\usage{
get_arg_min_conditional(lhs, rhs)
}
\arguments{
\item{lhs}{\code{vector}, LHS}

\item{rhs}{\code{vector}, RHS}
}
\value{
\code{saclar}, index of rhs[i] / lhs[i] for lhs[i] > 0
}
\description{
get_arg_min_conditional
}
