% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_simplex_algorithm.R
\name{get_dual_pivot_column}
\alias{get_dual_pivot_column}
\title{get_dual_pivot_colum}
\usage{
get_dual_pivot_column(t, dual_pivot_row)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}

\item{dual_pivot_row}{\code{scalar}, index of the pivot row}
}
\value{
\code{scalar}, index of the pivot column
}
\description{
get_dual_pivot_colum
}
