% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_simplex_algorithm.R
\name{get_dual_pivot_row}
\alias{get_dual_pivot_row}
\title{get_dual_pivot_row}
\usage{
get_dual_pivot_row(t)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}
}
\value{
\code{scalar}, index of the pivot row
}
\description{
get_dual_pivot_row
}
