% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{get_pivot_column}
\alias{get_pivot_column}
\title{get_pivot_column}
\usage{
get_pivot_column(t)
}
\arguments{
\item{t}{a Simplex tableau}
}
\value{
pivot column
}
\description{
get_pivot_column
}
