% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_simplex_simplex.R
\name{get_pivot_column_Bland}
\alias{get_pivot_column_Bland}
\title{get_pivot_column_Bland}
\usage{
get_pivot_column_Bland(t)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}
}
\value{
\code{scalar}, index of the pivot column
}
\description{
Choose the lowest-numbered (i.e., leftmost) nonbasic column with a negative (reduced) cost.
}
