% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_simplex_algorithm.R
\name{is_dual_optimal}
\alias{is_dual_optimal}
\title{is_dual_optimal}
\usage{
is_dual_optimal(t)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}
}
\value{
\code{logical}, TRUE = tableau is (dual) optimal, FALSE = tableau is not (dual) optimal
}
\description{
is_dual_optimal
}
