% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{pivot}
\alias{pivot}
\title{pivot}
\usage{
pivot(t, pivot_row, pivot_column)
}
\arguments{
\item{t}{a Simplex tableau}

\item{pivot_row}{the pivot row}

\item{pivot_column}{the pivot column}
}
\value{
new Simplex tableau
}
\description{
pivot
}
