% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepares The 2nd phase Table}
\usage{
prepare(mat, z_row, drop)
}
\arguments{
\item{mat}{A matrix representing a tableau}

\item{z_row}{A vetor representing Z row (1st row) of 2nd phase\preformatted{Note that this function does replace the Z row
at first and then drop the columns of artificial
variable. So Z_row vector length should be same
as the previous Z row of previous tableau matrix.
}}

\item{drop}{A vector giving the column indices of artificial variable to remove.}
}
\value{
A matrix representing the tableau for next step.
}
\description{
prepare the 2nd phase tablue in two-phase simplex method by replacing Z-row
}
\examples{
# please see the package vignette with \code{vignette("simplex", package = "simplexR")}
}
