% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revised_primal_simplex_agorithm.R
\name{revised_primal_simplex}
\alias{revised_primal_simplex}
\title{revised_primal_simplex}
\usage{
revised_primal_simplex(A, b, obj, basic = c((ncol(A) - nrow(A)):ncol(A)))
}
\arguments{
\item{A}{\code{matrix}, coefficient matrix of the LP}

\item{b}{\code{vector}, RHS}

\item{obj}{\code{vector}, coefficients of the objective function}

\item{basic}{\code{vector}, indices of a (primal feasible) basis}
}
\value{
optimal basic feasible solution (BFS)
}
\description{
Revised primal simplex algorithm.
}
