% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{simplex}
\alias{simplex}
\title{simplex}
\usage{
simplex(A, b, c, sense)
}
\arguments{
\item{A}{a matrix (coefficient matrix)}

\item{b}{a vector, RHS}

\item{c}{a vector, objective function coefficients}

\item{sense}{max = 1}
}
\value{
optimal Simplex tableau or stops execution in case of an unbounded problem
}
\description{
simplex
}
