% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_simplex_simplex.R
\name{simplexR}
\alias{simplexR}
\title{simplexR}
\usage{
simplexR(A, b, c, sense = 1, relation = rep("<=", length(b)), max_iter = 100)
}
\arguments{
\item{A}{\code{matrix}, coefficient matrix of the LP}

\item{b}{\code{vector}, RHS}

\item{c}{\code{vector}, coefficients of the objective function}

\item{sense}{\code{scalar}, max = 1, min = -1 (default = 1)}

\item{relation}{\code{vector}, {"<=", "=", "=>"} (default = "<=")}

\item{max_iter}{\code{scalar}, maximum number of iterations (default = 100)}
}
\value{
Solves (find an optimal solution or reports that the problem is not solveable) a generic linear program. Calls \code{construct_tableau} and \code{simplex}.
}
\description{
Solves (find an optimal solution or reports that the problem is not solveable) a generic linear program. Calls \code{construct_tableau} and \code{simplex}.
}
