% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau}
\alias{tableau}
\title{Calculates simplex tableau}
\usage{
tableau(mat, col, row)
}
\arguments{
\item{mat}{A matrix representing a tableau}

\item{col}{An integer representing pivot column of current tableau}

\item{row}{An integer representing pivot row of current tableau}
}
\value{
A matrix representing the tableau for next step
}
\description{
Calculates simplex tableau given a matrix and pivot column and pivot row of that matrix
}
\examples{
# Maximize  Z = 4x1 + 3x2 + 6x3 Subject to, 3x1 + x2 + 3x3 <= 30
# 2x1 + 2x2 + 3x3 <= 40 where x1 >= 0, x2 >= 0, x3 >= 0.

# Define the linear programming problem by a augmented matrix form.
prob_mat <- matrix(c(-4, -3, -6, 0, 0, 0, 3, 1, 3, 1,
                      0, 30, 2, 2, 3, 0, 1, 40),
                     byrow = TRUE, ncol = 6)

# check how that matrix looks like and find pivot row and pivot column.
prob_mat

# pivot column is 3rd one and pivot row is 2nd one.
prob_mat <- tableau(prob_mat, col = 3, row = 2)

# now pivot column is 2nd one and pivot row is 3rd one.
prob_mat <- tableau(prob_mat, col = 2, row = 3)
}
