% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_elim.R
\name{z_elim}
\alias{z_elim}
\title{Prepares the Z row}
\usage{
z_elim(mat, col, row)
}
\arguments{
\item{mat}{A matrix representing a tableau.}

\item{col}{An integer representing column which contains non-zero basic variable
coefficients in Z-row.}

\item{row}{An integer representing row by which you will eliminate the Z-row's
non-zero basic varible's coefficients.}
}
\value{
A matrix representing the tableau for next step.
}
\description{
Eliminates Basic Variables from the Z row of simplex tableau.
}
\examples{
# please see the package vignette with \code{vignette("simplex", package = "simplexR")}
}
