% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{construct_tableau}
\alias{construct_tableau}
\title{construct_tableau}
\usage{
construct_tableau(A, b, c, sense)
}
\arguments{
\item{A}{a matrix, the coefficient matrix of the LP}

\item{b}{a vector, RHS}

\item{c}{a vector, objective function coefficients}

\item{sense}{max = 1}
}
\value{
a matrix that represents the initial Simplax tableau
}
\description{
Constrcut an initial Simplex tableau of the linear program in canonical form for given A, b, c, sense (max = 1)
}
