% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{construct_tableau}
\alias{construct_tableau}
\title{construct_tableau}
\usage{
construct_tableau(
  A,
  b,
  c,
  sense = 1,
  relation = rep("<=", length(b)),
  bigM = 1000
)
}
\arguments{
\item{A}{\code{matrix}, coefficient matrix of the LP}

\item{b}{\code{vector}, RHS}

\item{c}{\code{vector}, coefficients of the objective function}

\item{sense}{\code{scalar}, max = 1, min = -1 (default = 1)}

\item{relation}{\code{vector}, {"<=", "=", "=>"} (default = "<=")}

\item{bigM}{\code{scalar}, M used in bigM-method (default = 1000)}
}
\value{
\code{matrix}, initial simplex tableau
}
\description{
Construct an initial simplex tableau of the linear program for given A, b, c, sense (max = 1, min = -1), relation ({"<=", "=", "=>"}).
}
