% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_simplex_functions.R
\name{get_pivot_column}
\alias{get_pivot_column}
\title{get_pivot_column}
\usage{
get_pivot_column(t)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}
}
\value{
\code{scalar}, index of the pivot column
}
\description{
Choose the first nonbasic column with the lowest negative (reduced) cost.
}
