% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{get_pivot_row}
\alias{get_pivot_row}
\title{get_pivot_row}
\usage{
get_pivot_row(t, pivot_column)
}
\arguments{
\item{t}{matrix, a Simplex tableau}

\item{pivot_column}{integer, the pivot column}
}
\value{
pivot row
}
\description{
get_pivot_row
}
