% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_simplex_simplex.R
\name{get_pivot_row}
\alias{get_pivot_row}
\title{get_pivot_row}
\usage{
get_pivot_row(t, pivot_column)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}

\item{pivot_column}{\code{scalar}, index of the pivot column}
}
\value{
\code{scalar}, index of the pivot row
}
\description{
get_pivot_row
}
