% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_simplex_functions.R
\name{optimality_check}
\alias{optimality_check}
\title{optimality_check}
\usage{
optimality_check(t)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}
}
\value{
\code{logical}, TRUE = tableau is optimal, FALSE = tableau is not optimal
}
\description{
optimality_check
}
