% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{optimality_check}
\alias{optimality_check}
\title{optimality_check}
\usage{
optimality_check(t)
}
\arguments{
\item{t}{\code{matrix}, a Simplex Tableau}
}
\value{
\code{logical}, (TRUE = optimal, FALSE = not optimal)
}
\description{
optimality_check
}
