% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{pivot}
\alias{pivot}
\title{pivot}
\usage{
pivot(t, pivot_row, pivot_column)
}
\arguments{
\item{t}{matrix, a Simplex tableau}

\item{pivot_row}{integer, the pivot row}

\item{pivot_column}{integer, the pivot column}
}
\value{
new Simplex tableau
}
\description{
pivot
}
