% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{pivot}
\alias{pivot}
\title{pivot}
\usage{
pivot(t, pivot_row, pivot_column)
}
\arguments{
\item{t}{\code{matrix}, a simplex tableau}

\item{pivot_row}{\code{scalar}, index of the pivot column}

\item{pivot_column}{\code{scalar}, index of the pivot row}
}
\value{
\code{matrix}, new simplex tableau
}
\description{
Performs a pivot operation an a given simplex tableau
}
