% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_functions.R
\name{simplex}
\alias{simplex}
\title{simplex}
\usage{
simplex(A, b, c, sense)
}
\arguments{
\item{A}{matrix, the coefficient matrix of the LP}

\item{b}{vector, RHS}

\item{c}{vector, objective function coefficients}

\item{sense}{integer, 1 = max}
}
\value{
optimal Simplex tableau or stops execution in case of an unbounded problem
}
\description{
simplex
}
