% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primal_simplex_functions.R
\name{simplex}
\alias{simplex}
\title{simplex}
\usage{
simplex(tableau, max_iter = 100)
}
\arguments{
\item{tableau}{\code{matrix}, simplex tableau in canonical form}

\item{max_iter}{\code{scalar}, maximum number of iterations}
}
\value{
optimal simplex tableau or stops execution in case of an unbounded problem
}
\description{
simplex
}
